working on userCategories' service
